Refactor `QrLoginHelpButton` to reusable `HelpButton`
